Got address to appear! For some reason, we call .call twice... Ohhhh, hold on, perhaps I just need to () it? Will test now
The point, though, is that the setItemFeatureAddress() state setter will allow us to re-render the page upon then Promise resolving, allowing it to appear smoothly! I suppose... the next step would be to check if defined, and if not, don't do anything - and rely on it being updated when it reloads?